requirements.txt

imgz-cli
pillow
PyYAML
